78 fiskalni račun (#79)
* fiskalizacija

* v4

* v4

* v5

* fiskalni v2

+++ app/src/main/java/com/example/roomapp/model/Product.kt
@@ -16,7 +16,7 @@ data class Product(
     val branchId: Int?,
     val deliveryStatus: String,
     val category: String?,
-    val price: Long
+    val price: Double
 ): Parcelable{
     override fun toString(): String {
         return prodName
